Use a template predicate for polling
NSPredicate(format:) will re-parse the format string every time the initializer is called. This is a simple predicate so it's not really that bad, but it's unnecessary. This way the predicate will be parsed once, and subsequent uses will re-use the original predicate with a substituted-in value for the latest timestamp